HYD-4387 simplify srcmap-reverse module
The current solution is hard to read, has way more code
than needed, and has more dependencies than are needed.
We are switching to highland streaming for our high
level flow control, and this module needs to be converted.
We need an integration test to test this library's side
of Joe's single line deminify capabililty
(in the Realtime service).
Also, we no longer want to use minijasminenodewrap,
but prefer jasmine-node.
In short, it needs to be pruned back for maintainability
and efficiency, and updated overall.

Bumped outdated libraries.
Dropped libraries that can be replaced with functional
composition.
Converted to highland streaming where appropriate.
Switched out minijasminenodewrap for jasmine-node.
Dropped almost 600 lines, and added less than 200 lines
of source code.

 - Replaced minijasminenodewrap with jasmine-node.
 - build-trace-collection.js  Refactored 87 lines to 20.
 - srcmap-reverse/index.js  Refactored 109 lines to 43.
 - Bumped version and dependencies, dropped minijasminenodewrap
   and promised-file.  Added highland, jasmine-node, and
   krusty-jasmine-reporter.
 - trace-reverse-test.js  Renamed from trace-reverse-spec.js and refactored
   from 80 to 61 lines.
 - trace-reverse.integration-test.js  Moved to functional composition, and
   added single line test.
 - Switched index.js to a HOF.

Signed-off-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>