Make repo a single package with cluster support Fixes #39.
This repo is currently setup as a monorepo using lerna and consists of
two packages (srcmap-reverse-server and srcmap-reverser). The point of
having it this way was to offload the process of sourcemapping
individual lines in a way that would scale. However, since we want to
bundle this as a single process that will run as a daemon it doesn't
make sense to publish two different bundles that must be run in the
background. Instead, this repo should be a single entity (without
lerna). A node cluster will be used to scale the processing of
individual lines.

- Integration test passing
- Integrated cluster to scale srcmap reverse process of individual lines
- Have a few more tests that I need to get passing but I wanted to get
  what I have committed.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>