GH-54 Update ExecStart location in srcmap-reverse.service (#55)
* GH-54 Update ExecStart location in srcmap-reverse.service
Fixes #GH-54.

We decided to move the service executable to /usr/lib/iml-srcmap-reverse. The ExecStart property needs to be updated
to look at this path in srcmap-reverse.service.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>

* GH-54 Remove unit files from repo
Fixes #54.

After a few discussions we've decided that the unit files are coupled to
the repo and should instead live in the dependencies repo. Delete the
unit files from this repo.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>